Time: 2152 ms (8.69%), Space: 262.7 MB (10.33%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -14,7 +14,7 @@ public:
                 a/=2;
             }
         }
-        long long val = 0,mul = 1;
+        int val = 0,mul = 1;
         for(int j=0;j<s.size();j++){
             val = (val+(s[j]-'0')*mul)%mod;
             mul*=2;
